% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{azure_store_anonymous}
\alias{azure_store_anonymous}
\title{Create an anonymous Azure store (unsigned).}
\usage{
azure_store_anonymous(account, container)
}
\arguments{
\item{account}{Storage account name.}

\item{container}{Blob container name.}
}
\value{
A \code{Store} object.
}
\description{
Azure does not generally permit anonymous listing even on "public"
containers — you will usually get a 401 \code{NoAuthenticationInformation}
error on \verb{store_list*()} calls. Anonymous \emph{reads} of known blob
names may still work depending on the container's public access
level. For most real Azure workflows use \code{\link[=azure_store_sas]{azure_store_sas()}} with
a SAS token, or \code{\link[=azure_store]{azure_store()}} with credentials from environment
variables.
}
