% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{local_store}
\alias{local_store}
\title{Create a local filesystem store rooted at \code{path}.}
\usage{
local_store(path)
}
\arguments{
\item{path}{Absolute or relative path to a directory.}
}
\value{
A \code{Store} object.
}
\description{
All keys passed to \code{put}/\code{get}/\code{list} are interpreted relative to \code{path}.
The directory must already exist.
}
