% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{memory_store}
\alias{memory_store}
\title{Create an in-memory store (useful for testing).}
\usage{
memory_store()
}
\value{
A \code{Store} object.
}
\description{
Create an in-memory store (useful for testing).
}
