% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{store_copy}
\alias{store_copy}
\title{Copy an object from \code{from} to \code{to} within the same store.}
\usage{
store_copy(store, from, to)
}
\arguments{
\item{store}{A \code{Store} object.}

\item{from}{Source key.}

\item{to}{Destination key.}
}
\value{
\code{NULL} invisibly on success.
}
\description{
Copy an object from \code{from} to \code{to} within the same store.
}
