% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{store_delete}
\alias{store_delete}
\title{Delete an object.}
\usage{
store_delete(store, key)
}
\arguments{
\item{store}{A \code{Store} object.}

\item{key}{Object key (path).}
}
\value{
\code{NULL} invisibly on success.
}
\description{
Delete an object.
}
