% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{store_exists}
\alias{store_exists}
\title{Check whether an object exists.}
\usage{
store_exists(store, key)
}
\arguments{
\item{store}{A \code{Store} object.}

\item{key}{Object key (path).}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Check whether an object exists.
}
