% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{store_get}
\alias{store_get}
\title{Read an entire object as a raw vector.}
\usage{
store_get(store, key)
}
\arguments{
\item{store}{A \code{Store} object.}

\item{key}{Object key (path).}
}
\value{
A raw vector containing the object's bytes.
}
\description{
Read an entire object as a raw vector.
}
