% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{store_put}
\alias{store_put}
\title{Write raw bytes to \code{key} in \code{store}.}
\usage{
store_put(store, key, data)
}
\arguments{
\item{store}{A \code{Store} object.}

\item{key}{Object key (path) as a string.}

\item{data}{A raw vector.}
}
\value{
\code{NULL} invisibly on success.
}
\description{
Write raw bytes to \code{key} in \code{store}.
}
